Made mana regeneration stop during casting and made the change also affect the mana costs of the spells

+++ src/arcane_arcade_field/WizardHudDrawer.java
@@ -366,7 +366,7 @@ public class WizardHudDrawer extends DrawableHandler
 		public void drawSelfBasic(Graphics2D g2d)
 		{
 			// Adjusts length before drawing the meter
-			setLength(WizardHudDrawer.this.wizard.getMana() / 10);
+			setLength((int) WizardHudDrawer.this.wizard.getMana() / 10);
 			
 			super.drawSelfBasic(g2d);
 		}
